feature/Connexion en tant que prof + ajout d'un profTest + vérifs

See merge request saedev/devapplicationformatiut!33

+++ src/vue/Admin/vueDetailEntreprise.php
@@ -29,11 +29,13 @@
 
         <div class="wrapBoutons">
             <?php
-            if ($entreprise->isEstValide()) {
-                echo '<a href="?action=supprimerEntreprise&controleur=AdminMain&siret=<?php echo $entreprise->getSiret() ?>">SUPPRIMER</a>';
-            } else {
-                echo '<a href="?action=refuserEntreprise&controleur=AdminMain&siret=' . $entreprise->getSiret() . '">REFUSER</a>';
-                echo '<a id="vert" href="?action=validerEntreprise&controleur=AdminMain&siret=' . $entreprise->getSiret() . '">ACCEPTER</a>';
+            if (\App\FormatIUT\Lib\ConnexionUtilisateur::getTypeConnecte()=="Administrateurs") {
+                if ($entreprise->isEstValide()) {
+                    echo '<a href="?action=supprimerEntreprise&controleur=AdminMain&siret=<?php echo $entreprise->getSiret() ?>">SUPPRIMER</a>';
+                } else {
+                    echo '<a href="?action=refuserEntreprise&controleur=AdminMain&siret=' . $entreprise->getSiret() . '">REFUSER</a>';
+                    echo '<a id="vert" href="?action=validerEntreprise&controleur=AdminMain&siret=' . $entreprise->getSiret() . '">ACCEPTER</a>';
+                }
             }
             ?>
         </div>
